Updated service to use extra transformer so the service itself is not plagued by unnecessary code
