+++ app/src/main/java/com/topjohnwu/magisk/di/ApplicationModule.kt
@@ -4,6 +4,8 @@ import android.content.Context
 import androidx.preference.PreferenceManager
 import com.skoumal.teanity.rxbus.RxBus
 import com.topjohnwu.magisk.App
+import com.topjohnwu.magisk.model.download.ModuleTransformer
+import com.topjohnwu.magisk.model.entity.internal.DownloadSubject
 import org.koin.dsl.module
 
 
@@ -15,4 +17,6 @@ val applicationModule = module {
     factory(Protected) { get<App>().protectedContext }
     single(SUTimeout) { get<Context>(Protected).getSharedPreferences("su_timeout", 0) }
     single { PreferenceManager.getDefaultSharedPreferences(get<Context>(Protected)) }
+
+    factory { (subject: DownloadSubject) -> ModuleTransformer(get(), subject) }
 }
